requirements.txt
transformers>=4.35.0
torch>=2.1.0
torchvision>=0.16.0
pillow>=10.0.0
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=4.9.0
supabase>=2.0.0
python-dotenv>=1.0.0
sentence-transformers>=0.5.0
transformers[torch]>=4.35.0
accelerate>=0.25.0